requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.5
djangorestframework==3.12.2
idna==2.10
Pillow==8.0.1
pytz==2020.1
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
